Move github client to its own module
Also defines a trait to allow easier testing without calling the actual
github apis.
