+++ Cargo.toml
@@ -7,9 +7,11 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+async-trait = "0.1.24"
 openssl-sys = { version = "0.9.54", features = ["vendored"] }
 reqwest = { version = "0.10.4", features = ["json"] }
 serde = { version = "1.0.104", features = ["derive"] }
+structopt = "0.3.11"
 tokio = { version = "0.2.13", features = ["macros"] }
 toml = "0.5.6"
-structopt = "0.3.11"
+
