+++ src/main.rs
@@ -1,11 +1,11 @@
+mod github;
+use github::{Github, GithubClient, Repository};
 use serde::Deserialize;
 use std::error::Error;
 use std::fs;
 use std::path::PathBuf;
 use structopt::StructOpt;
 
-static GRAM_USER_AGENT: &str = "gram";
-static GITHUB_BASE_URL: &str = "https://api.github.com";
 static SETTINGS_HELP: &str = concat!(
     "Path to the settings file",
     r#"
@@ -39,11 +39,6 @@ enum GramOpt {
 }
 
 #[derive(Debug, Deserialize)]
-struct Repository {
-    description: Option<String>,
-}
-
-#[derive(Debug, Deserialize)]
 struct Settings {
     description: Option<String>,
 }
@@ -69,32 +64,6 @@ fn read_settings(settings_location: &PathBuf) -> Result<Settings, Box<dyn Error>
     let settings_str = fs::read_to_string(settings_location)?;
     let settings = toml::from_str::<Settings>(&settings_str)?;
     Ok(settings)
-}
-
-struct Github {
-    client: reqwest::Client,
-}
-
-impl Github {
-    fn new() -> Self {
-        let client = reqwest::Client::builder()
-            .user_agent(GRAM_USER_AGENT)
-            .build()
-            .unwrap();
-        Self { client }
-    }
-
-    async fn repository(&self, owner: &str, repo: &str) -> Result<Repository, Box<dyn Error>> {
-        let repository = self
-            .client
-            .get(&format!("{}/repos/{}/{}", GITHUB_BASE_URL, owner, repo))
-            .header(reqwest::header::CONTENT_TYPE, "application/json")
-            .send()
-            .await?
-            .json::<Repository>()
-            .await?;
-        Ok(repository)
-    }
 }
 
 fn diff(repo: Repository, settings: Settings) -> Result<(), Box<dyn Error>> {
